Rename overview package to discovery
