+++ src/main/java/io/cryostat/platform/PlatformModule.java
@@ -51,17 +51,17 @@ import io.cryostat.core.sys.Environment;
 import io.cryostat.core.sys.FileSystem;
 import io.cryostat.messaging.notifications.NotificationFactory;
 import io.cryostat.net.AuthManager;
+import io.cryostat.platform.discovery.PlatformDiscoveryModule;
 import io.cryostat.platform.internal.CustomTargetPlatformClient;
 import io.cryostat.platform.internal.MergingPlatformClient;
 import io.cryostat.platform.internal.PlatformDetectionStrategy;
 import io.cryostat.platform.internal.PlatformStrategyModule;
-import io.cryostat.platform.overview.PlatformOverviewModule;
 
 import com.google.gson.Gson;
 import dagger.Module;
 import dagger.Provides;
 
-@Module(includes = {PlatformStrategyModule.class, PlatformOverviewModule.class})
+@Module(includes = {PlatformStrategyModule.class, PlatformDiscoveryModule.class})
 public abstract class PlatformModule {
 
     static final String PLATFORM_STRATEGY_ENV_VAR = "CRYOSTAT_PLATFORM";
