+++ src/main/java/io/cryostat/platform/discovery/BaseNodeType.java
@@ -35,46 +35,24 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.cryostat.platform.overview;
+package io.cryostat.platform.discovery;
 
-import java.util.Collections;
-import java.util.Map;
+public enum BaseNodeType implements NodeType {
+    // represents the entire deployment scenario Cryostat finds itself in
+    UNIVERSE("Universe"),
+    // represents a division of the deployment scenario - the universe may consist of a
+    // Kubernetes Realm and a JDP Realm, for example
+    REALM("Realm"),
+    ;
 
-import io.cryostat.platform.ServiceRef;
+    private final String kind;
 
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-
-public class TargetNode extends AbstractNode {
-    @SuppressFBWarnings("URF_UNREAD_FIELD")
-    private final ServiceRef target;
-
-    public TargetNode(NodeType nodeType, ServiceRef target) {
-        super(target.getServiceUri().toString(), nodeType, Collections.emptyMap());
-        this.target = target;
-    }
-
-    public TargetNode(NodeType nodeType, ServiceRef target, Map<String, String> labels) {
-        super(target.getServiceUri().toString(), nodeType, labels);
-        this.target = target;
+    BaseNodeType(String kind) {
+        this.kind = kind;
     }
 
     @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = super.hashCode();
-        result = prime * result + ((target == null) ? 0 : target.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) return true;
-        if (!super.equals(obj)) return false;
-        if (getClass() != obj.getClass()) return false;
-        TargetNode other = (TargetNode) obj;
-        if (target == null) {
-            if (other.target != null) return false;
-        } else if (!target.equals(other.target)) return false;
-        return true;
+    public String getKind() {
+        return kind;
     }
 }
