+++ src/main/java/io/cryostat/platform/discovery/CustomTargetNodeTypeAdapter.java
@@ -35,72 +35,34 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package io.cryostat.platform.overview;
+package io.cryostat.platform.discovery;
 
-import java.util.Collections;
-import java.util.Map;
+import java.io.IOException;
 
-import com.google.gson.annotations.SerializedName;
+import io.cryostat.platform.internal.CustomTargetPlatformClient;
+import io.cryostat.platform.internal.CustomTargetPlatformClient.CustomTargetNodeType;
+import io.cryostat.util.PluggableTypeAdapter;
 
-public abstract class AbstractNode implements Comparable<AbstractNode> {
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
 
-    protected final String name;
+public class CustomTargetNodeTypeAdapter extends PluggableTypeAdapter<CustomTargetNodeType> {
 
-    @SerializedName("kind")
-    protected final NodeType nodeType;
-
-    protected final Map<String, String> labels;
-
-    protected AbstractNode(String name, NodeType nodeType, Map<String, String> labels) {
-        this.name = name;
-        this.nodeType = nodeType;
-        this.labels = labels;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public NodeType getNodeType() {
-        return this.nodeType;
-    }
-
-    public Map<String, String> getLabels() {
-        return Collections.unmodifiableMap(labels);
+    public CustomTargetNodeTypeAdapter() {
+        super(CustomTargetNodeType.class);
     }
 
     @Override
-    public int compareTo(AbstractNode other) {
-        int type = nodeType.ordinal() - other.nodeType.ordinal();
-        if (type != 0) {
-            return type;
+    public CustomTargetNodeType read(JsonReader reader) throws IOException {
+        String token = reader.nextString();
+        if (CustomTargetPlatformClient.NODE_TYPE.getKind().equals(token)) {
+            return CustomTargetPlatformClient.NODE_TYPE;
         }
-        return name.compareTo(other.name);
+        return null;
     }
 
     @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((labels == null) ? 0 : labels.hashCode());
-        result = prime * result + ((name == null) ? 0 : name.hashCode());
-        result = prime * result + ((nodeType == null) ? 0 : nodeType.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) return true;
-        if (obj == null) return false;
-        if (getClass() != obj.getClass()) return false;
-        AbstractNode other = (AbstractNode) obj;
-        if (labels == null) {
-            if (other.labels != null) return false;
-        } else if (!labels.equals(other.labels)) return false;
-        if (name == null) {
-            if (other.name != null) return false;
-        } else if (!name.equals(other.name)) return false;
-        if (nodeType != other.nodeType) return false;
-        return true;
+    public void write(JsonWriter writer, CustomTargetNodeType nodeType) throws IOException {
+        writer.value(nodeType.getKind());
     }
 }
