+++ src/main/java/io/cryostat/platform/internal/KubeApiPlatformClient.java
@@ -52,10 +52,10 @@ import io.cryostat.core.net.JFRConnectionToolkit;
 import io.cryostat.core.net.discovery.JvmDiscoveryClient.EventKind;
 import io.cryostat.platform.ServiceRef;
 import io.cryostat.platform.ServiceRef.AnnotationKey;
-import io.cryostat.platform.overview.BaseNodeType;
-import io.cryostat.platform.overview.EnvironmentNode;
-import io.cryostat.platform.overview.NodeType;
-import io.cryostat.platform.overview.TargetNode;
+import io.cryostat.platform.discovery.BaseNodeType;
+import io.cryostat.platform.discovery.EnvironmentNode;
+import io.cryostat.platform.discovery.NodeType;
+import io.cryostat.platform.discovery.TargetNode;
 import io.cryostat.util.URIUtil;
 
 import dagger.Lazy;
